requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.12.1
drf-flex-fields==0.8.6
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
